requirements.txt
psycopg2-binary
pymongo
pandas
psutil
tqdm 